app/backend/api_server/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
atopile
